# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the main server where wallet transactions for the Más Simple Wallet will be received",
        "title": "Más Simple Wallet-Controller API",
        "contact": {
            "name": "API Support",
            "email": "ebasualdo@itba.edu.ar"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Returns a jwt token to use as authentication",
                "produces": [
                    "application/json"
                ],
                "summary": "Generate a jwt",
                "parameters": [
                    {
                        "description": "user's phone number and the received sms code",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.LoginDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dtos.TokenDto"
                        }
                    },
                    "401": {
                        "description": "Invalid phone and code combination"
                    }
                }
            }
        },
        "/auth/sms-code": {
            "post": {
                "description": "Sends an sms to the specified phoneNumber",
                "summary": "SMS auth",
                "parameters": [
                    {
                        "description": "user's phone number",
                        "name": "login",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.phoneNumberDto"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "400": {
                        "description": "Invalid phone number"
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        },
        "/me": {
            "get": {
                "description": "Returns an account by its id",
                "produces": [
                    "application/json"
                ],
                "summary": "Get account details",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Account"
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "description": "Replaces all the account information with the information pased",
                "produces": [
                    "application/json"
                ],
                "summary": "Edit account information",
                "parameters": [
                    {
                        "description": "Fields to edit",
                        "name": "account",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.AccountDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/me/instruments": {
            "get": {
                "description": "Returns a list of the available instruments uploaded by the client",
                "produces": [
                    "application/json"
                ],
                "summary": "Get available Instruments",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Instrument"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            },
            "post": {
                "description": "Inserts and instrument to the list of available user instruments\nReturn the instrument object with its id",
                "produces": [
                    "application/json"
                ],
                "summary": "Insert instrument",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the user that requests the instruments",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Instrument to insert",
                        "name": "instrument",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.InstrumentDto"
                        }
                    }
                ],
                "responses": {
                    "200": {},
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/me/transactions": {
            "get": {
                "description": "Returns all time transactions history",
                "produces": [
                    "application/json"
                ],
                "summary": "Get transaction history",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Transaction"
                            }
                        }
                    },
                    "404": {
                        "description": "No such user",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Executes as transaction and returns it's full details",
                "produces": [
                    "application/json"
                ],
                "summary": "execute transaction",
                "parameters": [
                    {
                        "description": "Transaction to execute to insert",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dtos.TransactionDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Transaction"
                        }
                    },
                    "400": {
                        "description": "The transaction provided is illegal",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dtos.AccountDto": {
            "type": "object",
            "required": [
                "dni",
                "email",
                "id",
                "lastName",
                "name",
                "phoneNumber"
            ],
            "properties": {
                "dni": {
                    "type": "string",
                    "example": "21399433"
                },
                "email": {
                    "type": "string",
                    "example": "moni.argento@hotmail.com"
                },
                "id": {
                    "type": "ID"
                },
                "instruments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.InstrumentDto"
                    }
                },
                "lastName": {
                    "type": "string",
                    "example": "Potrelli de Argento"
                },
                "name": {
                    "type": "string",
                    "example": "Mónica"
                },
                "phoneNumber": {
                    "type": "PhoneNumber"
                },
                "prepaids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dtos.PrepaidDto"
                    }
                }
            }
        },
        "dtos.InstrumentDto": {
            "type": "object",
            "required": [
                "creditType",
                "holder",
                "issuer",
                "lastFourNumbers",
                "pps",
                "validThru"
            ],
            "properties": {
                "creditType": {
                    "type": "string",
                    "enum": [
                        "DEBIT",
                        " CREDIT",
                        " PREPAID"
                    ],
                    "example": "DEBIT"
                },
                "holder": {
                    "type": "string",
                    "example": "José Pepe Argento"
                },
                "issuer": {
                    "type": "string",
                    "example": "Banco Itaú"
                },
                "lastFourNumbers": {
                    "type": "string",
                    "example": "4930"
                },
                "pps": {
                    "type": "string",
                    "enum": [
                        "VISA",
                        " AMEX",
                        " MC"
                    ],
                    "example": "VISA"
                },
                "validThru": {
                    "type": "string",
                    "example": "11/24"
                }
            }
        },
        "dtos.LoginDto": {
            "type": "object",
            "required": [
                "code",
                "phoneNumber"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "example": "123654"
                },
                "phoneNumber": {
                    "type": "string",
                    "example": "005491133071114"
                }
            }
        },
        "dtos.PrepaidDto": {
            "type": "object",
            "properties": {
                "QUILMES": {
                    "type": "string",
                    "example": "4200"
                },
                "availableDollar": {
                    "type": "number",
                    "example": 12.34
                },
                "availableDollarInAdvanced": {
                    "type": "number",
                    "example": 12.34
                },
                "availableLocal": {
                    "type": "number",
                    "example": 230.45
                },
                "availableLocalInAdvanced": {
                    "type": "number",
                    "example": 230.45
                },
                "creditDollar": {
                    "type": "number",
                    "example": -23.54
                },
                "creditLocal": {
                    "type": "number",
                    "example": -2850.63
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "dtos.TokenDto": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "expire": {
                    "type": "string",
                    "example": "2020-07-08T15:58:45+02:00"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJleHAiOjE1OTQyMTY3MjUsIm9yaWdfaWF0IjoxNTk0MjEzMTI1fQ.tWsDdREGVc2dPW7ZrcsoastWqfZm0s0w-oy6w0jH7YI"
                }
            }
        },
        "dtos.TransactionDto": {
            "type": "object",
            "required": [
                "amount",
                "destinationAccountId",
                "instrumentId"
            ],
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 1504.56
                },
                "destinationAccountId": {
                    "type": "ID"
                },
                "instrumentId": {
                    "type": "ID"
                }
            }
        },
        "handlers.phoneNumberDto": {
            "type": "object",
            "required": [
                "phoneNumber"
            ],
            "properties": {
                "phoneNumber": {
                    "type": "string",
                    "example": "005491133071114"
                }
            }
        },
        "models.Account": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-07T11:38:09.157803072Z"
                },
                "dni": {
                    "type": "string",
                    "example": "21399433"
                },
                "email": {
                    "type": "string",
                    "example": "moni.argento@hotmail.com"
                },
                "id": {
                    "type": "string"
                },
                "instruments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instrument"
                    }
                },
                "lastName": {
                    "type": "string",
                    "example": "Potrelli de Argento"
                },
                "name": {
                    "type": "string",
                    "example": "Mónica"
                },
                "phoneNumber": {
                    "type": "string",
                    "example": "005491133071114"
                },
                "prepaids": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Prepaid"
                    }
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                }
            }
        },
        "models.Instrument": {
            "type": "object",
            "properties": {
                "accountId": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-07 13:36:15.738848+02:00"
                },
                "creditType": {
                    "type": "string",
                    "enum": [
                        "DEBIT",
                        " CREDIT",
                        " PREPAID"
                    ],
                    "example": "DEBIT"
                },
                "holder": {
                    "type": "string",
                    "example": "José Pepe Argento"
                },
                "id": {
                    "type": "string"
                },
                "issuer": {
                    "type": "string",
                    "example": "Banco Itaú"
                },
                "lastFourNumbers": {
                    "type": "string",
                    "example": "4930"
                },
                "pps": {
                    "type": "string",
                    "enum": [
                        "VISA",
                        " AMEX",
                        " MC"
                    ],
                    "example": "VISA"
                },
                "validThru": {
                    "type": "string",
                    "example": "11/24"
                }
            }
        },
        "models.Prepaid": {
            "type": "object",
            "properties": {
                "QUILMES": {
                    "type": "string",
                    "example": "4200"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number",
                    "example": 1504.56
                },
                "createdAt": {
                    "type": "string",
                    "example": "2020-07-07 13:36:15.738848+02:00"
                },
                "destinationAccountId": {
                    "type": "string",
                    "example": "19fdsasfads"
                },
                "id": {
                    "type": "string"
                },
                "instrumentId": {
                    "type": "string",
                    "example": "asdfafasdfad5"
                },
                "originAccountId": {
                    "type": "string",
                    "example": "asdfasdfasdf3"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWT-Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}